yt-studio/requirements.txt
click==8.1.3
colorama==0.4.4
Flask==2.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.1.0
Werkzeug==2.1.2
zipp==3.8.0
